[test] Deduplicate `http_server_port`
